data_architecture/aic_project/requirements.txt
httplib2==0.19.1
pandas==1.3.2
mysql.connector==2.2.9
sqlalchemy==1.4.23
